requirements.txt
# --- Core Framework & Utilities ---
gradio>=4.0.0
pyyaml
piexif
pillow
tqdm
huggingface_hub>=0.24.3

# --- Major ML Libraries ---
# Note: User prefers to install torch manually, so we include it as a comment or broad spec
# torch>=2.1.0
torchvision
transformers>=4.43.3
accelerate>=0.26.0
safetensors

# --- Shared Model Dependencies ---
timm
einops
sentencepiece
protobuf
bitsandbytes
scipy

# --- Model Specific Dependencies ---

# ToriiGate / Qwen2-VL
qwen_vl_utils
# flash-attn --no-build-isolation  # Qwen3-VL: Flash Attention 2 support (manual install required)
# PyQt5 (optional, used by some ToriiGate variants)

# JoyCaption / Llama
# liger_kernel # Linux only (Triton dependency)

# Pixtral / JoyTag
tiktoken

# SmolVLM2
num2words

# Miscellaneous
# spaces # HuggingFace Spaces specific helper
colorama

# WD14 Tagger
onnxruntime-gpu
opencv-python-headless